Added question titles and changed answers to all be yes
Added in titles before the question to test if this makes the questions
more reusable if the caseworkers can scan for the question they’re
looking for.

Changed all question answers to yes opening the more details panel to
test if this speeds up ability to answer questions repeatedly if you
can scan the title and answer without having to re-read to see if it
needs yes or no.